+++ pennylane/ops/op_math/__init__.py
@@ -73,6 +73,8 @@ Decompositions
 .. autosummary::
     :toctree: api
 
+    ~one_qubit_decomposition
+    ~two_qubit_decomposition
     ~ctrl_decomp_zyz
     ~ctrl_decomp_bisect
 
@@ -83,6 +85,7 @@ from .composite import CompositeOp
 from .condition import cond, Conditional
 from .controlled import Controlled, ControlledOp, ctrl
 from .controlled_ops import ControlledQubitUnitary, CY, CZ
+from .decompositions import one_qubit_decomposition, two_qubit_decomposition, sk_decomposition
 from .evolution import Evolution
 from .exp import Exp, exp
 from .pow import Pow, pow
